requirements.txt
discord.py==1.4.1
PyNaCl==1.3.0
ffmpeg==1.4
discord.py[voice]
youtube-dl==2020.9.14
dnspython==1.16.0
cffi==1.14.0
wheel
gspread
oauth2client
